requirements.txt
torch==2.7.0
torchvision
zennit-crp
pytorch-lightning
pandas
opencv-python
wandb
scikit-learn
umap-learn
timm
safetensors
transformers
diffusers
open_clip_torch
einops
accelerate
dash-bootstrap-components
notebook
beautifulsoup4
scikit-image
streamlit
numpy==1.26.4
matplotlib==3.7.0
huggingface_hub
jaxtyping
fancy_einsum
line_profiler
kaleido
openai-clip
